requirements.txt
python-telegram-bot==20.6
flask
playwright
pandas
openpyxl
lxml